// frontend/src/api/models/ModelsBook.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Some API
 * This is a sample server celler server.
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@swagger.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ModelsBook
 */
export interface ModelsBook {
    /**
     * 
     * @type {string}
     * @memberof ModelsBook
     */
    author: string;
    /**
     * 
     * @type {string}
     * @memberof ModelsBook
     */
    finishedAt?: string;
    /**
     * 
     * @type {string}
     * @memberof ModelsBook
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof ModelsBook
     */
    purchasedAt?: string;
    /**
     * 
     * @type {number}
     * @memberof ModelsBook
     */
    rating?: number;
    /**
     * 
     * @type {string}
     * @memberof ModelsBook
     */
    title: string;
}

/**
 * Check if a given object implements the ModelsBook interface.
 */
export function instanceOfModelsBook(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "author" in value;
    isInstance = isInstance && "id" in value;
    isInstance = isInstance && "title" in value;

    return isInstance;
}

export function ModelsBookFromJSON(json: any): ModelsBook {
    return ModelsBookFromJSONTyped(json, false);
}

export function ModelsBookFromJSONTyped(json: any, ignoreDiscriminator: boolean): ModelsBook {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'author': json['author'],
        'finishedAt': !exists(json, 'finishedAt') ? undefined : json['finishedAt'],
        'id': json['id'],
        'purchasedAt': !exists(json, 'purchasedAt') ? undefined : json['purchasedAt'],
        'rating': !exists(json, 'rating') ? undefined : json['rating'],
        'title': json['title'],
    };
}

export function ModelsBookToJSON(value?: ModelsBook | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'author': value.author,
        'finishedAt': value.finishedAt,
        'id': value.id,
        'purchasedAt': value.purchasedAt,
        'rating': value.rating,
        'title': value.title,
    };
}